refactor: separate UIMessageChunk conversion from model streaming
Extract UIMessageChunk conversion from doStreamStep into a standalone
utility (to-ui-message-chunk.ts), making the model streaming layer
independent of UI concerns.

Key changes:
- doStreamStep no longer takes a writable parameter — it returns raw
  LanguageModelV4StreamPart chunks alongside tool calls, finish, and
  step result
- New toUIMessageChunk() pure function and createUIMessageChunkTransform()
  TransformStream for converting raw stream parts to UIMessageChunks
- New writeChunksToUI() step function in stream-text-iterator handles
  piping raw chunks through the UI transform to the writable
- writable is now optional in DurableAgentStreamOptions — when omitted,
  the agent streams ModelMessages only without UIMessageChunk conversion
- All UIMessageChunk writing (tool results, finish, close) is conditional
  on writable being provided

This follows the same pattern as streamText's toUIMessageStream() where
raw stream parts are the primary output and UI conversion is a separate,
optional step.
