+++ packages/durable-agent/src/stream-text-iterator.test.ts
@@ -117,7 +117,7 @@ function createMockDoStreamStepResult({
       finishReason: finishUnified,
       ...stepOverrides,
     }),
-    uiChunks: undefined,
+    chunks: [],
     providerExecutedToolResults: new Map(),
   };
 }
@@ -158,7 +158,7 @@ describe('streamTextIterator', () => {
           }),
         )
         .mockImplementationOnce(
-          async (prompt, _modelInit, _writable, _tools, _options) => {
+          async (prompt, _modelInit, _tools, _options) => {
             // Capture the prompt on the second call to verify providerOptions
             capturedPrompt = prompt;
             return createMockDoStreamStepResult();
@@ -240,7 +240,7 @@ describe('streamTextIterator', () => {
           }),
         )
         .mockImplementationOnce(
-          async (prompt, _modelInit, _writable, _tools, _options) => {
+          async (prompt, _modelInit, _tools, _options) => {
             capturedPrompt = prompt;
             return createMockDoStreamStepResult();
           },
@@ -319,7 +319,7 @@ describe('streamTextIterator', () => {
           }),
         )
         .mockImplementationOnce(
-          async (prompt, _modelInit, _writable, _tools, _options) => {
+          async (prompt, _modelInit, _tools, _options) => {
             capturedPrompt = prompt;
             return createMockDoStreamStepResult();
           },
@@ -422,7 +422,7 @@ describe('streamTextIterator', () => {
           }),
         )
         .mockImplementationOnce(
-          async (prompt, _modelInit, _writable, _tools, _options) => {
+          async (prompt, _modelInit, _tools, _options) => {
             capturedPrompt = prompt;
             return createMockDoStreamStepResult();
           },
@@ -511,7 +511,7 @@ describe('streamTextIterator', () => {
           }),
         )
         .mockImplementationOnce(
-          async (prompt, _modelInit, _writable, _tools, _options) => {
+          async (prompt, _modelInit, _tools, _options) => {
             capturedPrompt = prompt;
             return createMockDoStreamStepResult();
           },
@@ -583,7 +583,7 @@ describe('streamTextIterator', () => {
           }),
         )
         .mockImplementationOnce(
-          async (prompt, _modelInit, _writable, _tools, _options) => {
+          async (prompt, _modelInit, _tools, _options) => {
             capturedPrompt = prompt;
             return createMockDoStreamStepResult();
           },
@@ -661,7 +661,7 @@ describe('streamTextIterator', () => {
           }),
         )
         .mockImplementationOnce(
-          async (prompt, _modelInit, _writable, _tools, _options) => {
+          async (prompt, _modelInit, _tools, _options) => {
             capturedPrompt = prompt;
             return createMockDoStreamStepResult();
           },
